requirements.txt
opencv-python-headless
pillow
tqdm
numpy
pandas
matplotlib
torchvision
tensorflow
ultralytics==8.0.180
streamlit
dill==0.3.5.1